backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client==1.16.1
python-dotenv==1.0.0
openai==1.3.5
psycopg[binary]==3.1.13
pydantic==2.5.2
python-multipart==0.0.6
google-generativeai>=0.8.3
